chap05_adding_json_fields/restapi_project/requirements.txt
# python 3.5.x
boto3==1.4.7
Django==1.11.5
django-cors-headers==2.1.0
django-rest-swagger==2.1.2
djangorestframework==3.6.4
gunicorn==19.7.1
httpie==0.9.9
whitenoise==3.3.1
